+++ src/jevdroid/android/__init__.py
@@ -1,3 +1,4 @@
+from jevdroid.android.adb import AdbDevice
 from jevdroid.android.device import AndroidDevice
 
-__all__ = ["AndroidDevice"]
+__all__ = ["AdbDevice", "AndroidDevice"]